requirements.txt
pick
gitpython
termcolor
